# api/openapi.yaml
openapi: 3.0.3
info:
  title: Karvon Scraper API
  version: 1.0.0
  description: |
    REST API for Karvon: the scraper module (Google Maps lead scraping with website
    email crawling) and the campaign module (outreach campaigns, contacts and their
    lifecycle, reusable email content, consent and the two delivery providers). Live
    job progress is delivered over Server-Sent Events.

    All routes except `/healthz`, `/openapi.json` and `/webhooks/*` require
    `Authorization: Bearer <KARVON_API_KEY>`. The `/webhooks/*` routes carry no API
    key: they are called by Instantly and Mailchimp, and are authenticated instead by
    the per-registration token in their path together with the secret that provider
    signs its deliveries with.
servers:
  - url: /api/v1
    description: Default mount point
security:
  - bearerAuth: []
...
paths:
...
  /jobs/{id}/events:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [jobs]
      operationId: streamJobEvents
      summary: Server-Sent Events stream of job progress, logs and status
      description: |
        `text/event-stream`. Event names are `progress`, `log` and `status`; every event
        carries a monotonic `id`. Reconnect with the `Last-Event-ID` header (or `?after=`)
        to replay missed events. A `: ping` comment is sent every 15 seconds. The stream
        closes once the job reaches a terminal status.
      parameters:
        - name: after
          in: query
          description: Replay events with an id greater than this value
          schema: { type: integer, format: int64, minimum: 0 }
        - name: Last-Event-ID
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Static API key, `Authorization: Bearer <KARVON_API_KEY>`"
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            details:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
      additionalProperties: false
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
...